Refactor global context
