+++ src/app/contexts/ExamplesContext/constants.ts
