+++ src/app/index.tsx
@@ -2,18 +2,18 @@ import React from 'react';
 
 import '../scss/application.scss';
 
-import { reducer, INITIAL_STATE, Action, GlobalState } from '~contexts/GlobalContext/reducer';
-import { Context } from '~app/contexts/GlobalContext';
+import { reducer, INITIAL_STATE, Action, ExamplesState } from '~app/contexts/ExamplesContext/reducer';
+import { Context } from '~app/contexts/ExamplesContext';
 import withProvider from '~components/ProviderWrapper';
 
 import Routes from './components/Routes';
-import { EXAMPLES_LOCAL_STORAGE_KEY } from './contexts/GlobalContext/constants';
+import { EXAMPLES_LOCAL_STORAGE_KEY } from './contexts/ExamplesContext/constants';
 
 function App() {
   return <Routes />;
 }
 
-export default withProvider<{}, GlobalState, Action>({
+export default withProvider<{}, ExamplesState, Action>({
   Context,
   reducer,
   initialState: INITIAL_STATE,
